Test 'repeated field annotations'
